backend/requirements.txt
fastapi
uvicorn
openai
PyPDF2
python-multipart
aiofiles
python-dotenv